+++ src/components/custom-field-input-date/custom-field-input-date.jsx
@@ -1,10 +1,13 @@
 import React, { forwardRef, useEffect, useImperativeHandle, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import calendarSvg from '../../svgs/calendar.svg';
+import IconButton from '../icon-button/icon-button.jsx';
 import Icon from '../icon/icon.jsx';
 import { convertToFormat, validDate } from './format/format-date.js';
-import styles from '../__internal__/abstract-custom-field/abstract-custom-field.css';
+import dateStyles from './custom-field-date.css';
 import AbstractCustomField from '../__internal__/abstract-custom-field/abstract-custom-field.jsx';
+import useDropdownClose from '../../hooks/use-dropdown-close.js';
+import Calendar from '../calendar/calendar.jsx';
 
 const isValidInput = (value) => {
   if (value === '' || value === undefined) {
@@ -13,6 +16,12 @@ const isValidInput = (value) => {
 
   return validDate(value);
 };
+
+function isValidDate(dateString) {
+  if (dateString === '') return true;
+  const date = new Date(`${dateString}T00:00`);
+  return date instanceof Date && !isNaN(date);
+}
 
 const isValueValid = (value, error, isInputValid = false) => {
   if (error) {
@@ -29,29 +38,33 @@ const isValueValid = (value, error, isInputValid = false) => {
 const CustomFieldInputDate = forwardRef(function CustomFieldInputDate(props, ref) {
   const componentRef = useRef(null);
   const inputRef = useRef(null);
-  const [isEditing, setIsEditing] = useState(true);
-  const [isValid, setIsValid] = useState(isValueValid(props.value, props.error, true));
-  const [isFocused] = useState(false);
+  const [isEditing, setIsEditing] = useState(false);
+  const value = validDate(props.value) ? convertToFormat(props.value, 'yyyy-mm-dd') : props.value;
+  const [currentValue, setCurrentValue] = useState(value);
+  const [isValid, setIsValid] = useState(isValueValid(currentValue, props.errorText, true));
+  const [expanded, setExpanded] = useState(false);
+  const [shouldFocusInput, setShouldFocusInput] = useState(false);
 
   useEffect(() => {
-    if (!inputRef.current) return;
-
-    const isInputValid = inputRef.current.validity.valid;
-    const valid = isValueValid(props.value, props.error, isInputValid);
-    setIsValid(valid);
-
-    if (!valid) {
-      setIsEditing(true);
-    } else {
-      setIsEditing(false);
-    }
-  }, [inputRef.current, props.value, props.error]);
-
-  useEffect(() => {
-    if (isEditing && isFocused) {
+    if (isEditing && inputRef.current) {
       inputRef.current.focus();
     }
-  }, [isEditing, isFocused]);
+  }, [isEditing]);
+
+  useEffect(() => {
+    if (shouldFocusInput && componentRef.current) {
+      componentRef.current.focus();
+    }
+
+    setShouldFocusInput(false);
+  }, [shouldFocusInput]);
+
+  function onDateSelected(date) {
+    setShouldFocusInput(true);
+    setCurrentValue(date.toISOString().slice(0, 10));
+    setExpanded(false);
+    setIsEditing(false);
+  }
 
   useImperativeHandle(ref, () => ({
     id: props.id,
@@ -64,9 +77,10 @@ const CustomFieldInputDate = forwardRef(function CustomFieldInputDate(props, ref
     if (inputRef.current) {
       const isInputValid = inputRef.current.validity.valid;
       const newDate = convertToFormat(event.target.value, 'yyyy-mm-dd');
-      setIsValid(isValueValid(newDate, props.error, isInputValid));
+      setIsValid(isValueValid(newDate, props.errorText, isInputValid));
     }
 
+    setCurrentValue(event.target.value);
     props.onChange(event);
   };
 
@@ -78,61 +92,136 @@ const CustomFieldInputDate = forwardRef(function CustomFieldInputDate(props, ref
     return props.errorText;
   };
 
-  const sharedProps = {
-    className: props.className,
-    disabled: props.disabled,
-    icon: <Icon className={styles['input-icon']} icon={calendarSvg} label={props.label} />,
-    label: props.label,
-    inputRef,
-    readOnly: true,
-    required: props.required,
+  const wrapperRef = useRef(null);
+  const handleDropdownClose = () => {
+    setExpanded(false);
+    setIsEditing(false);
+  };
+  useDropdownClose(wrapperRef, expanded, handleDropdownClose);
+
+  const openCalendar = () => {
+    if (!props.disabled) {
+      setExpanded(!expanded);
+      setIsEditing(!isEditing);
+    }
   };
 
-  if (isEditing || !isValid) {
-    const value = validDate(props.value) ? convertToFormat(props.value, 'yyyy-mm-dd') : props.value;
-
-    return (<AbstractCustomField
-      {...sharedProps}
-      defaultValue={value}
-      errorText={errorText()}
-      id={props.id}
-      key={`${props.id}-editing`}
-      min={convertToFormat(props.min, 'yyyy-mm-dd')}
-      max={convertToFormat(props.max, 'yyyy-mm-dd')}
-      onChange={handleOnChange}
-      step={1}
-      type="date"
-    />);
+  function onInputClick(event) {
+    event.preventDefault();
+    if (!props.disabled) {
+      setExpanded(true);
+      setIsEditing(true);
+    }
   }
 
-  return (<AbstractCustomField
-    {...sharedProps}
-    defaultValue={convertToFormat(props.value, 'Month dd, yyyy')}
-    id={props.id}
-    key={`${props.id}-readonly`}
-    inputRef={componentRef}
-    type="text"
-  />);
+  function onInputKeyDown(event) {
+    if (event.key === 'Enter' || event.key === 'Space') {
+      event.preventDefault();
+      if (!props.disabled) {
+        setShouldFocusInput(true);
+        setExpanded(!expanded);
+        setIsEditing(!isEditing);
+      }
+    }
+  }
+
+  function onKeyDown(event) {
+    if (event.key === 'Escape') {
+      setExpanded(false);
+      setIsEditing(false);
+    }
+  }
+
+  function calendarIcon() {
+    if (props.disabled) {
+      return (
+        <Icon
+          icon={calendarSvg}
+          title={props.label}
+          label={`${props.label} calendar`}
+        />
+      );
+    }
+
+    return (
+      <IconButton
+        onPress={openCalendar}
+        icon={calendarSvg}
+        title={props.label}
+        label={`${props.label} calendar button`}
+      />
+    );
+  }
+
+  const sharedProps = {
+    className: dateStyles['date-input'],
+    inputClassName: props.errorText ? dateStyles['input-invalid'] : dateStyles.input,
+    disabled: props.disabled,
+    icon: calendarIcon(),
+    label: props.label,
+    inputRef,
+    required: props.required,
+    onClick: onInputClick,
+    onKeyDown: onInputKeyDown,
+    errorText: props.errorText,
+  };
+
+  function renderField() {
+    if (isEditing || !isValid) {
+      return (
+        <AbstractCustomField
+          {...sharedProps}
+          defaultValue={currentValue}
+          errorText={errorText()}
+          id={props.id}
+          key={`${props.id}-editing`}
+          min={convertToFormat(props.min, 'yyyy-mm-dd')}
+          max={convertToFormat(props.max, 'yyyy-mm-dd')}
+          onChange={handleOnChange}
+          step={1}
+          type="date"
+        />);
+    }
+
+    return (
+      <AbstractCustomField
+        {...sharedProps}
+        defaultValue={convertToFormat(currentValue, 'Month dd, yyyy')}
+        id={props.id}
+        key={`${props.id}-readonly`}
+        inputRef={componentRef}
+        type="text"
+      />
+    );
+  }
+
+  return (
+    // eslint-disable-next-line jsx-a11y/no-static-element-interactions
+    <div ref={wrapperRef} onKeyDown={onKeyDown}>
+      { renderField() }
+      { expanded && isValidDate(currentValue) && (
+        <div className={dateStyles['calendar-container']}>
+          <Calendar value={currentValue} onDateSelected={onDateSelected} />
+        </div>
+      )}
+    </div>
+  );
 });
 
 CustomFieldInputDate.propTypes = {
-  className: PropTypes.string,
   disabled: PropTypes.bool,
-  error: PropTypes.bool,
   errorText: PropTypes.string,
   id: PropTypes.string.isRequired,
   label: PropTypes.string.isRequired,
-  min: PropTypes.string,
-  max: PropTypes.string,
+  min: PropTypes.string, /* eslint-disable-line react/no-unused-prop-types */
+  max: PropTypes.string, /* eslint-disable-line react/no-unused-prop-types */
   onChange: PropTypes.func,
   required: PropTypes.bool,
   value: PropTypes.string,
 };
 
 CustomFieldInputDate.defaultProps = {
-  className: undefined,
   disabled: false,
-  error: false,
   errorText: '',
   min: undefined,
   max: undefined,